Update isort config to use_parentheses instead of combine_as_imports (#547)
The `combine_as_imports=True` modifies isort style as a side-effect and was not the intended purpose of the suggested change in #250. The problem was that isort was actually replacing the parens with backslash and using `combine_as_imports=True` happened to also produce the same result.

The actual setting should be `use_parentheses` as this tells isort to use parenthesis for line continuation instead of \ for lines over the allotted line length limit and matches precisely what black is outputting.

(cherry picked from commit 158f796ef3ac0b41d39d8674aa08199d338b30d9)